requirements.txt
cfgv==3.3.1
distlib==0.3.4
filelock==3.4.2
identify==2.4.1
lxml==4.7.1
nodeenv==1.6.0
platformdirs==2.4.1
pre-commit==2.16.0
python-docx==0.8.11
PyYAML==6.0
six==1.16.0
toml==0.10.2
virtualenv==20.11.2
